Converted the .NET 8 demo app to MVVM using the Microsoft CommunityToolkit.MVVM package
