+++ DemoApps/DemoAppNet8/App.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System.Globalization;
+﻿using Microsoft.Extensions.DependencyInjection;
+using System.Globalization;
 using System.Windows;
 using WPFLocalizeExtension.Engine;
 
@@ -8,7 +9,25 @@ namespace DemoAppNet8
     {
         public App()
         {
+            Services = ConfigureServices();
+
             LocalizeDictionary.Instance.Culture = new CultureInfo("en");
+
+            InitializeComponent();
+        }
+
+        public new static App Current => (App)Application.Current;
+
+        public IServiceProvider Services { get; }
+
+        private static IServiceProvider ConfigureServices()
+        {
+            var services = new ServiceCollection();
+
+            // ViewModels
+            services.AddTransient<MainWindowViewModel>();
+
+            return services.BuildServiceProvider();
         }
     }
 }
